Replaced jsonwebtoken with jose for test tokens
jsonwebtoken internally import `utils` which is no
longer available in a browser environment in Cypress
15.x.

We use it just to create test JWT tokens in a normal test and
in integration tests.

This may not be ideal as I had to resort to dynamically
import jose as Jest was complaining about the module
format used by it. This also means passing --experimental-vm-modules
to jest.
