+++ package.json
@@ -21,8 +21,8 @@
     "start-feature:dev": "node esbuild/esbuild --watch --env feature.env",
     "lint": "eslint . --cache --max-warnings 0",
     "typecheck": "tsc && tsc -p integration_tests",
-    "test": "jest",
-    "test:ci": "jest --detectOpenHandles",
+    "test": "NODE_OPTIONS=\"$NODE_OPTIONS --experimental-vm-modules\" jest",
+    "test:ci": "NODE_OPTIONS=\"$NODE_OPTIONS --experimental-vm-modules\" jest --detectOpenHandles",
     "security_audit": "npx audit-ci --config audit-ci.json",
     "int-test": "cypress run",
     "int-test-ui": "cypress open --e2e",
@@ -79,7 +79,6 @@
     "@types/http-errors": "^2.0.5",
     "@types/jest": "^30.0.0",
     "@types/jquery": "^4.0.0",
-    "@types/jsonwebtoken": "^9.0.10",
     "@types/node": "^24.12.2",
     "@types/nunjucks": "^3.2.6",
     "@types/nunjucks-date": "^0.0.10",
@@ -101,7 +100,7 @@
     "jest": "^30.3.0",
     "jest-html-reporter": "^4.4.0",
     "jest-junit": "^16.0.0",
-    "jsonwebtoken": "^9.0.3",
+    "jose": "6.2.2",
     "lint-staged": "^16.4.0",
     "mocha-junit-reporter": "^2.2.1",
     "nock": "^14.0.12",
